# internal/api/api.swagger.yaml
swagger: '2.0'
info:
  title: Space VPX Backend Service
  version: development
  description: Space VPX WebSite Backend Service
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /solar_panel_top:
    post:
      security:
        - Bearer: []
      summary: Create Solar Panel Top
      operationId: CreateSolarPanelTop
      tags: 
        - SolarPanelTop
      parameters:
        - name: CreateSolarPanelTopBody
          description: Create Solar Panel Top Body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CreateSolarPanelTopBody'
      responses:
        200:
          description: Create Solar Panel Top Response
          schema:
            $ref: '#/definitions/SolarPanelTop'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Forbidden
          schema:
            $ref: '#/definitions/Error' 
        422:
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/Error"        
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
  /solar_panel_top/available_solar_panel_top:
    get:
      security:
        - Bearer: []
      summary: Get Cube Sat Solar Panels Top
      operationId: GetCubeSatSolarPanelsTop
      tags: 
        - SolarPanelTop
      parameters:
      - in: query
        required: true
        name: offset
        type: integer
        description: Offset Configs
        minimum: 0
        required: true
      - in: query
        required: true
        name: limit
        type: integer
        description: Offset Configs
        minimum: 0
        required: true
      - in: query
        name: sort[field]
        description: sort parameters
        enum: [created_at]
        type: string
      - in: query
        name: FilterCubeSatSolarPanelByLength[min]
        description: Filter By Min Length
        type: number
      - in: query
        name: FilterCubeSatSolarPanelByLength[max]
        description: Filter By Max Length
        type: number
      responses:
        200:
          description: Get CubeSat Solar Panels Top Response
          schema:
            $ref: '#/definitions/CubeSatSolarPanelsTop'
        400:
          description: Bad request
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Forbidden
          schema:
            $ref: '#/definitions/Error' 
        422:
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/Error"        
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'


###SOLARPANELSIDE
...
definitions:
...
  Error:
    type: object
    required:
      - message
    properties:
      message:
        type: string
...
  SolarPanelTop:
    type: object
    required:
      - id
    properties:
      id:
        type: integer
        format: int64
        description: Unique identifier for the solar panel top (auto-generated)
      name:
        type: string
      length:
        type: number
        description: Length of the solar panel top
      width:
        type: number
        description: Width of the solar panel top
      height:
        type: number
        description: Height of the solar panel top
      max_operating_temperature:
        type: number
        description: Maximum operating temperature in degrees Celsius
      min_operating_temperature:
        type: number
        description: Minimum operating temperature in degrees Celsius
      weight:
        type: number
        description: Weight of the solar panel top
      interface:
        type: string
        description: Interface type of the solar panel top
      voc:
        type: number
        description: Open circuit voltage
      isc:
        type: number
        description: Short circuit current
      vmp:
        type: number
        description: Voltage at maximum power
      imp:
        type: number
        description: Current at maximum power
      efficiency:
        type: number
        description: Efficiency percentage
      coil_area:
        type: number
        description: Area of the coil
      coil_resistance:
        type: number
        description: Coil resistance value
      mechanical_vibration:
        type: number
        description: Mechanical vibration value
      mechanical_shock:
        type: number
        description: Mechanical shock value
      updated_at:
        type: integer
        description: Timestamp of the last update
      created_at:
        type: integer
        description: Timestamp of creation
  CubeSatSolarPanelsTop:
    type: object
    required:
      - count
    properties:
      count:
        type: integer
      cube_sat_solar_panels_top:
        type: array
        items:
          $ref: "#/definitions/SolarPanelTop"
  CreateSolarPanelTopBody:
    type: object
    required:
      - length
      - width
      - height
    properties:
      name:
        type: string
      length:
        type: number
      width:
        type: number
      height:
        type: number
      max_operating_temperature:
        type: number
      min_operating_temperature:
        type: number
      weight:
        type: number
      interface:
        type: string
      voc:
        type: number
      isc:
        type: number
      vmp:
        type: number
      imp:
        type: number
      efficiency:
        type: number
      coil_area:
        type: number
      coil_resistance:
        type: number
      mechanical_vibration:
        type: number
      mechanical_shock:
        type: number
...